🔧 Improve reminder
