+++ server/src/client/src/components/shared/Progress/Progress.module.scss
@@ -1,3 +1,10 @@
+:root {
+    --bg-secondary: #f0f0f0;
+    --color-primary: #4f46e5;
+    --color-success: #2ecc71;
+    --color-danger: #e74c3c;
+}
+
 .progressContainer {
   width: 100%;
   height: 6px;
